Macro moveref::bind

macro_rules! bind {
    (mut $name:ident $(: $ty:ty)? = &move *$expr:expr) => { ... };
    ($name:ident $(: $ty:ty)? = &move *$expr:expr) => { ... };
    (mut $name:ident $(: $ty:ty)? = &move $expr:expr) => { ... };
    ($name:ident $(: $ty:ty)? = &move $expr:expr) => { ... };
    (mut $name:ident $(: $ty:ty)? = $expr:expr) => { ... };
    ($name:ident $(: $ty:ty)? = $expr:expr) => { ... };
    (@move $(($mut:tt))? $name:ident, $($ty:ty)?, $expr:expr) => { ... };
    (@put $(($mut:tt))? $name:ident, $($ty:ty)?, $expr:expr) => { ... };
    (@emplace $(($mut:tt))? $name:ident, $($ty:ty)?, $expr:expr) => { ... };
}

Macro for binding a variable to a fresh MoveRef.

bind!(x = &move *ptr) creates an x: MoveRef<T> given ptr: impl (DerefMove + DerefMut<Target = T>)

The above invocation moves the referent of a uniquely owning poiner into a fresh MoveRef bound to x.

bind!(x = &move val) creates an x: MoveRef<T> given val: T

The above invocation moves any value into a fresh MoveRef bound to x.

bind!(x = con) creates an x: Pin<MoveRef<T>> given con: impl New<Output = T>

The above invocaton constructs a New value into a fresh MoveRef bound to x.

bind!(mut x: T = ...) (with right-hand side of &move *ptr or &move val or con)

The above generalization can be used with any earlier invocation form to add mutability and typing annotations.