Macro moveref::expr

macro_rules! expr {
    (&move *$expr:expr) => { ... };
    (&move $expr:expr) => { ... };
    ($expr:expr) => { ... };
}

Macro for creating a fresh MoveRef expression.

Because a v: MoveRef<'frame, T> always has an associated lifetime 'frame, this macro can generally only be used where it would be immediately consumed by some enclosing expression, such as in the position of a function argument.

Trying to use this as let v = expr!(...) will not work because the lifetime 'frame will not expand to the enclosing let binding. Hence the need for the separate bind!(v = ...) macro.

Otherwise, the usage is generally the same as bind!(...):

expr!(&move *ptr) creates a MoveRef<T> given ptr: impl (DerefMove + DerefMut<Target = T>)

The above invocation moves the referent of a uniquely owning poiner into a fresh MoveRef.

expr!(&move val) creates an MoveRef<T> given val: T

The above invocation moves any value into a fresh MoveRef.

expr!(con) creates an x: Pin<MoveRef<T>> given con: impl New<Output = T>

The above invocaton constructs a New value into a fresh MoveRef.