Trait moveref::new::TryNew

pub trait TryNew {
    type Output;
    type Error;

    // Required method
    unsafe fn try_new(
        self,
        this: Pin<&mut MaybeUninit<Self::Output>>
    ) -> Result<(), Self::Error>;
}

Types which can be constructed (initialized) into some provided storage. Construction may fail.

Required Associated Types

type Output

type Error

Required Methods

unsafe fn try_new( self, this: Pin<&mut MaybeUninit<Self::Output>> ) -> Result<(), Self::Error>

Try to initialize Self using this for storage.

Errors

Should return Err if initialization failed.

Safety

TryNew::try_new() must not be used to mutate previously initialized data

this must be freshly-allocated memory

after invocation, the this placement argument is in a valid, initialized state