Trait moveref::DerefMove

pub unsafe trait DerefMove: DerefMut + IntoMove {
    // Required method
    fn deref_move<'frame>(
        self,
        storage: Slot<'frame, Self::Storage>
    ) -> MoveRef<'frame, Self::Target>
       where Self: 'frame;
}

Safety

Correctness for DerefMove impls require that the uniqueness invariant of MoveRef is upheld. In particular, the following function must not violate memory safety:

fn move_out_of<P>(p: P) -> P::Target
where
  P: DerefMove,
  P::Target: Sized,
{
  unsafe {
    // Replace `p` with a move reference into it.
    bind!(p = &move *p);

    // Move out of `p`. From this point on, the `P::Target` destructor must
    // run when, and only when, the function's return value goes out of
    // scope per the usual Rust rules.
    //
    // In particular, the original `p` or any pointer it came from must not
    // run the destructor when they go out of scope, under any circumstance.
    MoveRef::into_inner(p)
  }
}

Required Methods

fn deref_move<'frame>( self, storage: Slot<'frame, Self::Storage> ) -> MoveRef<'frame, Self::Target>where Self: 'frame,